+++ src/lib.rs
@@ -65,7 +65,7 @@
 //! assert_eq!(problem.domain(), &"briefcase-world".into());
 //! assert!(problem.requirements().is_empty());
 //! assert_eq!(problem.init().len(), 9);
-//! assert!(matches! { problem.goal(), pddl::PreGD::And(_) });
+//! assert!(matches! { problem.goal(), pddl::PreconditionGoalDefinition::And(_) });
 //! ```
 
 // only enables the `doc_cfg` feature when